partition synthesis and place-and-route flow properly.
